Rename init -> data_initialization
